+++ user-service/src/main/java/com/sobok/userservice/common/security/JwtFilter.java
@@ -37,7 +37,7 @@ public class JwtFilter extends OncePerRequestFilter {
     private final ObjectMapper objectMapper;
 
     List<String> whiteList = List.of(
-            "/actuator/**", "/user/findByPhoneNumber", "/auth/findLoginId", "/api/**"
+            "/actuator/**", "/user/findByPhoneNumber", "/api/**"
     );
 
     @Override
