+++ user-service/src/main/java/com/sobok/userservice/common/security/SecurityConfig.java
@@ -32,7 +32,7 @@ public class SecurityConfig {
                 // 허용 URI 설정
                 .authorizeHttpRequests(auth -> auth
                         .requestMatchers(
-                                "/actuator/**", "/user/findByPhoneNumber", "/auth/findLoginId", "/api/**"
+                                "/actuator/**", "/auth/findLoginId", "/api/**"
                         ).permitAll()
                         .anyRequest().authenticated()
                 )
