+++ user-service/src/main/java/com/sobok/userservice/user/controller/UserController.java
@@ -23,16 +23,6 @@ public class UserController {
     private final UserService userService;
     private final UserAddressService userAddressService;
 
-    @PostMapping("/findByPhoneNumber")
-    public ResponseEntity<?> getUser(@RequestBody String phoneNumber) {
-
-        UserResDto byPhoneNumber = userService.findByPhoneNumber(phoneNumber);
-
-        log.info("검색한 사용자 정보 with phone number: {}", byPhoneNumber);
-
-        return ResponseEntity.ok().body(ApiResponse.ok(byPhoneNumber, "전화번호로 찾은 user 정보입니다."));
-
-    }
 
     @PostMapping("/addAddress")
     public ResponseEntity<?> addAddress(@AuthenticationPrincipal TokenUserInfo userInfo, @RequestBody UserAddressReqDto reqDto) {
